+++ SharedData/Models/AmazonLaptopProcessedModel.cs
@@ -4,7 +4,7 @@ public class AmazonLaptopProcessedModel
 {
     public string? Brand { get; set; }
     public double? ScreenSize { get; set; }
-    public int? Ram { get; set; }
-    public int? HardDisk { get; set; }
+    public double? Ram { get; set; }
+    public double? HardDisk { get; set; }
     public double? Price { get; set; }
 }
